+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix removing hidden lines of code blocks that are quoted
+- Fix removing hidden lines of code blocks that are quoted, in a list, etc.
 
 ## [0.18.0] - 2025-08-30